Edit: Separate navbar component + css module file
